removed the static watcher, made npm watch tasks
- renamed and relocated files to make more sense
- file renames and README update
