+++ lib/developer-help.js
