+++ package.json
@@ -16,26 +16,33 @@
     "node": "^4.4.1"
   },
   "scripts": {
-    "app": "npm run prep && node app.js",
-    "build": "npm run prep",
-    "copy-static-files": "run-s copy:img copy:vendor",
+    "cleanup": "shx rm -rf build dist",
     "copy:img": "shx mkdir -p dist/img && shx cp -r img/* dist/img",
-    "copy:vendor": "run-s copy:vendor:bootstrap copy:vendor:font-awesome copy:vendor:tabzilla copy:vendor:mofo-ui",
     "copy:vendor:bootstrap": "shx mkdir -p dist/vendor/bootstrap && shx cp -r node_modules/bootstrap/dist/* dist/vendor/bootstrap",
     "copy:vendor:font-awesome": "shx mkdir -p dist/vendor/font-awesome && shx cp -r node_modules/font-awesome/* dist/vendor/font-awesome",
-    "copy:vendor:tabzilla": "shx mkdir -p dist/vendor/font-awesome && shx cp node_modules/mozilla-tabzilla/css/tabzilla.css dist/vendor/font-awesome && shx cp -r node_modules/mozilla-tabzilla/media/img dist/vendor/font-awesome",
+    "copy:vendor:tabzilla": "shx mkdir -p dist/vendor/mozilla-tabzilla/css && shx mkdir -p dist/vendor/mozilla-tabzilla/media/img && shx cp node_modules/mozilla-tabzilla/css/tabzilla.css dist/vendor/mozilla-tabzilla/css && shx cp -r node_modules/mozilla-tabzilla/media/img/* dist/vendor/mozilla-tabzilla/media/img",
     "copy:vendor:mofo-ui": "shx mkdir -p dist/vendor/mofo-ui && shx cp node_modules/mofo-ui/dist/mofo-ui.css dist/vendor/mofo-ui",
-    "cleanup": "shx rm -rf dist",
+    "copy:vendor:locale-data": "shx mkdir -p dist/vendor/react-intl/locale-data && shx cp node_modules/react-intl/locale-data/* dist/vendor/react-intl/locale-data",
+    "copy-static-files": "run-s copy:img copy:vendor:*",
     "l10n": "pontoon-to-json",
     "less": "lessc less/index.less dist/styles.css --source-map=dist/styles.css.map --clean-css=\"--s0\"",
-    "lint": "eslint --ext .js --ext .jsx .",
-    "lint:fix": "npm run lint -- --fix",
-    "prep": "run-s cleanup copy-static-files lint less l10n webpack",
-    "postinstall": "npm run l10n",
-    "start": "npm run app",
-    "test": "run-s prep",
-    "webpack": "webpack --config ./config/webpack.config.js --progress --colors",
-    "webpack:dev": "webpack-dev-server --config ./config/webpack.config.js --progress --colors --watch"
+    "lint:fix": "eslint --ext .js --ext .jsx . -- --fix",
+    "postinstall": "npm run prep",
+    "prep": "run-s cleanup copy-static-files less l10n webpack:static webpack",
+    "start": "run-p start:*",
+    "start:less": "chokidar-cmd -c 'npm run less' -t less",
+    "start:app": "npm run webpack -- --watch",
+    "start:static": "npm run webpack:static -- --watch",
+    "start:server": "node app",
+    "test": "npm run prep",
+    "webpack": "webpack --config ./config/webpack/webpack.client.config.js --progress --colors",
+    "webpack:static": "webpack --config ./config/webpack/webpack.server.config.js --progress --colors"
+  },
+  "babel": {
+    "presets": [
+      "es2015",
+      "react"
+    ]
   },
   "dependencies": {
     "babel-cli": "^6.11.4",
@@ -45,9 +52,13 @@
     "babel-preset-react": "^6.11.1",
     "bootstrap": "^3.3.4",
     "chalk": "^1.1.1",
+    "chokidar": "^1.6.0",
+    "chokidar-cmd": "^1.2.1",
     "classnames": "^2.2.1",
     "comment-parser": "^0.3.1",
     "es5-shim": "^4.4.1",
+    "eslint": "^3.3.1",
+    "eslint-loader": "^1.5.0",
     "esprima-harmony-jscs": "^1.1.0-bin",
     "express": "^4.12.3",
     "font-awesome": "^4.3.0",
@@ -59,6 +70,8 @@
     "intl-locales-supported": "^1.0.0",
     "json-loader": "^0.5.1",
     "leaflet.markercluster": "^0.4.0",
+    "less": "^2.7.1",
+    "less-plugin-clean-css": "^1.5.1",
     "mapbox.js": "^2.1.5",
     "merge-stream": "^1.0.0",
     "minimatch": "^3.0.0",
@@ -84,6 +97,7 @@
     "react-router": "^1.0.3",
     "react-select": "^0.9.1",
     "react-wp-content-loader": "0.0.2",
+    "require-uncached": "^1.0.2",
     "routington": "^1.0.3",
     "should": "^8.0.2",
     "shx": "^0.1.4",
@@ -101,9 +115,6 @@
     "webpack-stream": "^3.1.0"
   },
   "devDependencies": {
-    "eslint": "^3.3.1",
-    "less": "^2.7.1",
-    "less-plugin-clean-css": "^1.5.1",
     "marked": "^0.3.3",
     "open": "0.0.5",
     "simplecrawler": "^0.5.2",
